"Moved models from backend/models.py to backend/__init__.py and removed backend/models.py"
