+++ backend/__init__.py
@@ -1,8 +1,8 @@
+from backend.models import Annotation, Project, Video, create_db_models, initialize_db
 from backend.utils import get_environment_variable, parse_bool
 
 from .application import create_app, run_application
 from .cognito_util import sign_up, verify_sign_up
-from .models import Annotation, Project, Video, create_db_models, initialize_db
 from .routes import (
     register_cognito_auth_endpoints,
     register_projects_endpoint,
