+++ backend/models/__init__.py
@@ -1,0 +1,4 @@
+from .annotation import Annotation
+from .database import create_db_models, initialize_db
+from .project import Project
+from .video import Video
